# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/google": {
      "post": {
        "operationId": "signInWithGoogle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signInWithGoogle_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SignInResult"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [],
        "summary": "Login or register with Google OAuth",
        "tags": ["auth"]
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "SignUpResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SignInResult": {
        "properties": {
          "account": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SignUpResult"
              }
            ]
          },
          "token": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "signInWithGoogle_request": {
        "properties": {
          "idToken": {
            "type": "string"
          }
        },
        "required": ["idToken"],
        "type": "object"
      },
...
    }
  }
}